transform UI > components, rename UI to theme
